fix tensorboard

+++ rl/core.py
@@ -208,9 +208,9 @@ class Agent(object):
 
                     # This episode is finished, report and reset.
                     episode_logs = {
-                        'episode_reward': episode_reward,
-                        'nb_episode_steps': episode_step,
-                        'nb_steps': self.step,
+                        'episode_reward': np.float_(episode_reward),
+                        'nb_episode_steps': np.float_(episode_step),
+                        'nb_steps': np.float_(self.step),
                     }
                     callbacks.on_episode_end(episode, episode_logs)
